fix: correct relative import paths in moved WebWrapper files
