+++ apps/web/modules/event-types/components/tabs/availability/EventAvailabilityTabWebWrapper.tsx
@@ -6,8 +6,8 @@ import { useLocale } from "@calcom/lib/hooks/useLocale";
 import type { RouterOutputs } from "@calcom/trpc/react";
 import { trpc } from "@calcom/trpc/react";
 
-import type { TeamMembers } from "../EventType";
-import { EventAvailabilityTab } from "../tabs/availability/EventAvailabilityTab";
+import type { TeamMembers } from "../../EventType";
+import { EventAvailabilityTab } from "./EventAvailabilityTab";
 
 export type EventAvailabilityTabWebWrapperProps = {
   eventType: EventTypeSetup;
